+++ core/index.js
@@ -2,4 +2,5 @@ module.exports = exports = {
 	terminal: require('./terminal'),
 	sockets: require('./sockets'),
 	workers: require('./workers'),
+	models: require('./models')
 };